backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
openpyxl
pandas
python-dotenv
aiofiles
pillow
httpx
